requirements.txt
openai
asyncio
cx-Freeze
LLMClassBasedToolRegistry
pyside6
python-dotenv
pre-commit
ruff
pytest-asyncio